Added PreviewLoaderTests
