+++ Example/PreFireExample.xcodeproj/xcshareddata/xcschemes/PrefireExample.xcscheme
@@ -38,16 +38,6 @@
                ReferencedContainer = "container:PrefireExample.xcodeproj">
             </BuildableReference>
          </TestableReference>
-         <TestableReference
-            skipped = "NO">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "PrefireExecutableTests"
-               BuildableName = "PrefireExecutableTests"
-               BlueprintName = "PrefireExecutableTests"
-               ReferencedContainer = "container:..">
-            </BuildableReference>
-         </TestableReference>
       </Testables>
    </TestAction>
    <LaunchAction
