+++ PrefireExecutable/Sources/prefire/Commands/Tests/GenerateTestsCommand.swift
