+++ PrefireExecutable/Sources/prefire/Extensions/Array+Args.swift
